Use new code boilerplate
See https://github.com/diamondburned/discord-workshop-env.
